\name{trim}
\alias{trim}
\title{
  This function is used to trim leading or trailing whitespaces
}
\description{
  This function is used to trim leading or trailing whitespaces
}
\usage{
  trim(x)
}

\arguments{
  \item{x}{
  Input String
}
}

\examples{
  trim("testString ")
}
